feat: 방 입장 시 비밀번호 및 인원 제한 검증 기능 추가

+++ src/chat/chat.service.ts
@@ -1,9 +1,11 @@
-import { Injectable, UnauthorizedException } from '@nestjs/common';
+import { BadRequestException, Injectable, UnauthorizedException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { Room } from 'src/rooms/rooms.entity';
 import { User } from 'src/users/users.entity';
 import { In, Repository } from 'typeorm';
 import { RedisService } from '../redis/redis.service';
 import { ChatEvents } from './chat.events';
+import * as bcrypt from 'bcrypt';
 
 @Injectable()
 export class ChatService {
@@ -11,23 +13,39 @@ export class ChatService {
     private readonly redisService: RedisService,
     @InjectRepository(User)
     private readonly userRepository: Repository<User>,
+    @InjectRepository(Room)
+    private readonly roomRepository: Repository<Room>,
     private readonly chatEvents: ChatEvents,
   ) {}
 
-  async joinRoom(roomId: number, userId: number) {
-    await this.redisService.addUserToRoom(roomId, userId);
+  async joinRoom(roomId: number, userId: number, password: any) {
+    const room = await this.roomRepository.findOne({ where: { id: roomId } });
+    if (!room) throw new BadRequestException('존재하지 않는 방입니다.');
+
+    const beforeCount = await this.redisService.getRoomUserCount(roomId);
+    if (beforeCount>=room.maxMembers) throw new BadRequestException('방의 인원이 가득 찼습니다.');
     
+    if (room.password) {
+      const isPasswordValid = await bcrypt.compare(password, room.password);
+      if (!isPasswordValid) throw new BadRequestException('비밀번호가 일치하지 않습니다.');
+    }
+
+    await this.redisService.addUserToRoom(roomId, userId);
+
+    const afterCount = await this.redisService.getRoomUserCount(roomId);
     const roomUsersArray = await this.redisService.getRoomUsers(roomId);
     const roomUsers = await this.userRepository.find({
       where: { id: In(roomUsersArray.map(userId => Number(userId))) },
       select: ['id','username']
     });
-    const roomUserCount = await this.redisService.getRoomUserCount(roomId);
 
-    return { roomUsers, roomUserCount }
+    return { roomUsers, afterCount }
   }
 
   async leaveRoom(roomId: number, userId: number) {
+    const isUserInRoom = await this.redisService.isUserInRoom(roomId, userId);
+    if (!isUserInRoom) throw new BadRequestException('방에 존재하지 않습니다.');
+
     await this.redisService.removeUserFromRoom(roomId, userId);
     const roomUsersArray = await this.redisService.getRoomUsers(roomId);
     const roomUsers = await this.userRepository.find({
@@ -35,6 +53,7 @@ export class ChatService {
       select: ['id','username']
     });
     const roomUserCount = await this.redisService.getRoomUserCount(roomId);
+    console.log('룸유저카운트',roomUserCount);
 
     return { roomUsers, roomUserCount }
   }
@@ -50,7 +69,6 @@ export class ChatService {
         const roomId = Number(key.split(':')[1]);
 
         const isUserInRoom = await this.redisService.isUserInRoom(roomId, session.userId);
-        
         if (!isUserInRoom) {
           console.log(`${roomId}번 방에는 ${session.username}님이 존재하지 않습니다.`);
           return;
